[TEST] Lleve a un archivo aparte la clase Arca
